+++ Assets/Scripts/DataTypes/GBC/LUDI/DataFile/LUDI_PocketPC_DataFile.cs
@@ -5,7 +5,7 @@ namespace R1Engine {
 		// Serialized properties
 		public LUDI_AppInfo AppInfo { get; set; }
 		public Pointer BaseOffset { get; set; }
-		public uint TotalLength { get; set; }
+		public long TotalLength { get; set; }
 
 		public override void SerializeImpl(SerializerObject s) {
 			AppInfo = s.SerializeObject<LUDI_AppInfo>(AppInfo, name: nameof(AppInfo));
@@ -32,13 +32,13 @@ namespace R1Engine {
 			return null;
 		}
 
-		public override uint? GetLength(ushort blockID) {
+		public override long? GetLength(ushort blockID) {
 			if(DataInfo != null) return DataInfo.DataSize + 4;
 			if (OffsetTable != null) {
 				if (!OffsetTable.EntriesDictionary.ContainsKey(blockID)) return null;
 				var entryIndex = OffsetTable.EntriesDictionary[blockID];
 				uint blockOffset = OffsetTable.Entries[entryIndex].BlockOffset;
-				uint nextBlockoffset = entryIndex < OffsetTable.Entries.Length - 1 ? OffsetTable.Entries[entryIndex+1].BlockOffset : TotalLength;
+				long nextBlockoffset = entryIndex < OffsetTable.Entries.Length - 1 ? OffsetTable.Entries[entryIndex+1].BlockOffset : TotalLength;
 				return nextBlockoffset - blockOffset;
 			}
 			return null;
